refactor: split inspect-run viewer seams
